requirements.txt
django
djangorestframework
mysqlclient
reportlab